Migrate navbar bs3/4 specific subclasses to build helpers and `base`
