+++ addon/components/bs3/bs-navbar.js
@@ -1,6 +1,1 @@
-import BaseNavbar from 'ember-bootstrap/components/base/bs-navbar';
-
-export default class Navbar extends BaseNavbar {
-  _positionPrefix = 'navbar-';
-  _validPositions = ['fixed-top', 'fixed-bottom', 'static-top'];
-}
+export { default } from 'ember-bootstrap/components/base/bs-navbar';
